[+]Добавлена текущая погода

+++ app/src/main/java/ru/korotaeva/vasilisa/weather2/model/weatherOfCity/Rain.kt
@@ -1,5 +1,5 @@
 package ru.korotaeva.vasilisa.weather2.model.weatherOfCity
 
 data class Rain(
-    val `1h`: Int
+    val `1h`: Double
 )
